pyproject.toml
[tool.poetry]
name = "langchain-sprint-project"
version = "0.1.0"
description = "Sprint's capstone project lanchain"
authors = ["pablo-lucas <pablo.lucv@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_sprint_project"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.273"
openai = "^0.27.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
colorlog = "^6.7.0"
pandas = "^2.1.0"
tabulate = "^0.9.0"
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
